pyproject.toml
[project]
name = "meeting-app-dec-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "dj-database-url>=2.3.0",
    "django>=6.0",
    "django-axes>=8.1.0",
    "django-crispy-forms>=2.5",
    "django-csp>=4.0",
    "django-htmx>=1.27.0",
    "gunicorn>=23.0.0",
    "nh3>=0.3.2",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dateutil>=2.9.0",
    "python-magic>=0.4.27",
    "weasyprint>=67.0",
    "whitenoise>=6.8.0",
]

[dependency-groups]
dev = [
    "bandit>=1.9.2",
]
